client: isolated unicode fix on Ubuntu
The fix is in particular for the following case:
- Running on linux
- Python process is started with LANG unset

In this case, fix_encoding.fix_encoding() fails to fix
sys.getfilesystemencoding() to be utf-8, and is instead set to 'ascii', which
means that non-ascii file paths fails to be mapped in. Swarming bots with LANG
correctly set do not exhibit this issue.

R=vadimsh@chromium.org
Bug: 787531
Change-Id: I7a754e1042ed0af7715e249546a7ceafaa269088
Reviewed-on: https://chromium-review.googlesource.com/807124
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Commit-Queue: Marc-Antoine Ruel <maruel@chromium.org>

Cr-Mirrored-From: https://chromium.googlesource.com/infra/luci/luci-py
Cr-Mirrored-Commit: 18c44b9f39caef0fdadaccba80b0d6fa998518b5