client/tools: add reproduce.py
This is for a very specific need: when someone needs to reproduce one or
multiples tasks that ran on a server on another server.

Generally it is to reproduce a task that ran on Production server to the Staging
server to test bot-side changes. But this can be generalized by tags, for
example running all Swarming tasks for a 'try job' on staging.

Not all flags (cipd, named cache, service account, env prefix) have been tested,
but it's still useful enough and the untested parts can be fixed later as
needed.

R=vadimsh@chromium.org
Bug:
Change-Id: I49d1ee4cd1ace5f9807ddfab4115996c9485128e
Reviewed-on: https://chromium-review.googlesource.com/806775
Commit-Queue: Marc-Antoine Ruel <maruel@chromium.org>
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>

Cr-Mirrored-From: https://chromium.googlesource.com/infra/luci/luci-py
Cr-Mirrored-Commit: b8e270dda0c3e91f98727c64c680c51c1688cce7